adding some compiler flags
