+++ github-watcher/src/main/scala/nhyne/package.scala
@@ -1,0 +1,7 @@
+import zio.logging.Logging
+import zio.ZEnv
+import nhyne.prometheus.Metrics.MetricsService
+
+package object nhyne {
+  type BaseEnv = Logging with MetricsService with ZEnv
+}
